♻️ refactor: fetcherを整理
