+++ src/repositories/config/fetchConfig.spec.ts
@@ -1,7 +1,7 @@
 import { mockConfig } from "mocks/data";
 import { server } from "mocks/msw/server";
 
-import { fetchConfig } from ".";
+import { fetchConfig } from "..";
 
 beforeAll(() => server.listen());
 afterAll(() => server.close());
